+++ src/App.java
@@ -15,7 +15,7 @@ public class App {
         String element[] = {"Fire", "Oven", "Sink", "Wall", "Washer", "Water"};
         Map map;
         try {
-            map = MapBuilder.generateFromFile("mappa.txt"); // generazione della mappa
+            map = MapBuilder.generateFromFile("mappa.txt", element); // generazione della mappa
             GUIView view = new GUIView(map); // generazione della Gui grafica in base alla struttura del modello
             TextualView view2 = new TextualView();
             GUIPartialView view3 = new GUIPartialView(map);
