+++ src/model/Cell.java
@@ -23,5 +23,7 @@ public abstract class Cell
         return j;
     }
 
+    public abstract char getC();
+
     public abstract javax.swing.ImageIcon getIcon();
 }
