+++ src/model/MapBuilder.java
@@ -2,7 +2,10 @@ package model;
 
 import java.io.File;
 import java.io.FileNotFoundException;
+import java.lang.reflect.Constructor;
 import java.util.Scanner;
+
+import javax.lang.model.util.ElementScanner6;
 
 import model.element.*;
 import model.exception.FileStructureWrongException;
@@ -10,7 +13,7 @@ import model.exception.MapToSmallException;
 
 public class MapBuilder {
 
-    public static Map generateFromFile(String fileName) throws FileNotFoundException, FileStructureWrongException{
+    public static Map generateFromFile(String fileName, String el) throws FileNotFoundException, FileStructureWrongException{
         File file = new File(fileName);
         Scanner myReader = new Scanner(file);
         int i;
@@ -21,42 +24,38 @@ public class MapBuilder {
             throw new FileStructureWrongException("Il file risulta vuoto");
         }
         try {
-            int MapSize = Integer.parseInt(myReader.nextLine());
-            Cell.Scale = MapSize;
-            mappa = new Cell[MapSize][MapSize];
-            for(i = 0; i < MapSize && myReader.hasNextLine(); i++)
+            int xMapSize = myReader.nextInt(), yMapSize = Integer.parseInt(myReader.nextLine());
+            mappa = new Cell[yMapSize][xMapSize];
+            for(i = 0; i < yMapSize && myReader.hasNextLine(); i++)
             {
-                //per ora va bene, ma non ci sono tutti gli elementi 
                 String data = myReader.nextLine();
-                for(int j = 0; j < data.length(); j+=2)
+                for(int j = 0, k = 0; j < xMapSize; j++, k+=2)
                 {
-                    switch (data.charAt(j)) {
-                        case 'W'://wall
-                            mappa[i][j/2] = new Wall(i, j/2);
-                            break;
-                        case 'S'://sink
-                            mappa[i][j/2] = new Sink(i, j/2);
-                            break;
-                        case 'w'://water
-                            mappa[i][j/2] = new Water(i, j/2);
-                            break;
-                        default:
-                            mappa[i][j/2] = null;
-                            break;
+                    final Constructor<Cell> c = CellBuilder.create(data.charAt(k), el);
+                    mappa[i][j] = c.newInstance(i, j);
+                
+                    if(c instanceof CellState)
+                    {
+                        k++;
+                        ((CellState)c).setState(data.charAt(k));
                     }
                 }
             }
 
-            if(i != MapSize)
+            if(i != yMapSize)
             {   myReader.close();
                 throw new FileStructureWrongException("La mappa è stata disegnata male");
             }
-        } catch (NumberFormatException e) {
-            myReader.close();
-
+        } 
+        catch (NumberFormatException e)
+        {
             throw new FileStructureWrongException(e.getMessage());
         }
-        myReader.close();
+        finally
+        {
+            myReader.close();
+        }
+        
         return new Map(mappa);
     }
 
@@ -65,10 +64,6 @@ public class MapBuilder {
         Cell mappa[][] = new Cell[10][10];
         Azzera(mappa);
         return new Map(mappa);
-    }
-
-    public static Map generateRandomMap(int size) throws MapToSmallException {
-        return generateRandomMap(size, size);
     }
 
     public static Map generateRandomMap(int xSize, int ySize) throws MapToSmallException {
@@ -82,30 +77,15 @@ public class MapBuilder {
     private static void Azzera(Cell[][] mappa) {
         for (int i = 0; i < mappa.length; i++) {
             for (int j = 0; j < mappa[i].length; j++) {
-                if (i == 0 || j == 0 || i == mappa.length - 1 || j == mappa[i].length - 1) {
+                //TODO sistema
+                /*if (i == 0 || j == 0 || i == mappa.length - 1 || j == mappa[i].length - 1) {
                     mappa[i][j] = new Wall(i, j);
                 } else {
                     mappa[i][j] = randCasella(i, j);
-                }
+                }*/
             }
         }
     }
 
-    private static Cell randCasella(int i, int j) {
-        double randomNum = Math.random();
-        randomNum = Math.round(randomNum * 100) / 1000;
-
-        if(randomNum < 0.1) {
-            return new Sink(i, j);
-        }
-        else if(randomNum < 0.2) {
-            return new Washer(i, j);
-        }
-        else if(randomNum < 0.3) {
-            return new Oven(i, j);
-        }
-        else{
-            return null;
-        }
-    }
+    
 }
